Move 'static-books.ts' into the 'data' directory
