+++ content/webapp/data/static-books.ts
